+++ data/rust/morloc-runtime-types/src/lib.rs
@@ -19,4 +19,5 @@ pub mod schema;
 pub mod cschema;
 pub mod null_check;
 pub mod packet;
+pub mod compression;
 pub mod daemon_socket;